News_Aggregator/requirements.txt
newspaper3k
spacy

